int-tests: Make it possible to start apim only once while writing tests. Add readme.
