+++ integration/test-integration/src/test/java/org/wso2/choreo/connect/tests/setup/withapim/WithApimBeforeTestSuite.java
@@ -19,84 +19,33 @@
 package org.wso2.choreo.connect.tests.setup.withapim;
 
 import org.awaitility.Awaitility;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeSuite;
-import org.wso2.am.integration.test.ClientAuthenticator;
-import org.wso2.am.integration.test.impl.RestAPIAdminImpl;
-import org.wso2.am.integration.test.impl.RestAPIPublisherImpl;
-import org.wso2.am.integration.test.impl.RestAPIStoreImpl;
-import org.wso2.am.integration.test.utils.base.APIMIntegrationConstants;
-import org.wso2.am.integration.test.utils.bean.DCRParamRequest;
-import org.wso2.choreo.connect.tests.apim.ApimAdvancedBaseTest;
 import org.wso2.choreo.connect.tests.context.ApimInstance;
 import org.wso2.choreo.connect.tests.util.HttpClientRequest;
 import org.wso2.choreo.connect.tests.util.HttpResponse;
 import org.wso2.choreo.connect.tests.util.Utils;
 
 import java.io.IOException;
-import java.net.URL;
 import java.util.Objects;
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
 /**
  * This class starts the API Manager instance before the entire test suite.
- * Here the class extends the ApimAdvancedBaseTest class in order to check whether
- * all the APIM REST clients can be initialized as expected when needed.
  */
-public class WithApimBeforeTestSuite extends ApimAdvancedBaseTest {
-    private static final Logger log = LoggerFactory.getLogger(WithApimBeforeTestSuite.class);
+public class WithApimBeforeTestSuite {
     ApimInstance apimInstance;
 
-    @BeforeSuite(description = "start API Manager and Prepare for the tests")
+    @BeforeSuite(description = "start API Manager")
     void startAPIM() throws Exception {
         apimInstance = new ApimInstance();
         apimInstance.startAPIM();
-
-        // set ssl properties needed to run the apim server. eg: truststore cert, truststore password, https protocols
-        setSSlSystemProperties();
-
-        Awaitility.await().pollDelay(2, TimeUnit.MINUTES).pollInterval(20, TimeUnit.SECONDS)
+        Awaitility.await().pollDelay(2, TimeUnit.MINUTES).pollInterval(15, TimeUnit.SECONDS)
                 .atMost(4, TimeUnit.MINUTES).until(isAPIMServerStarted());
-
-        String dcrURL = Utils.getAPIMServiceURLHttps("/client-registration/v0.17/register");
-        //DCR call for publisher app
-        DCRParamRequest publisherParamRequest = new DCRParamRequest(RestAPIPublisherImpl.appName,
-                                                                    RestAPIPublisherImpl.callBackURL,
-                                                                    RestAPIPublisherImpl.tokenScope,
-                                                                    RestAPIPublisherImpl.appOwner,
-                                                                    RestAPIPublisherImpl.grantType, dcrURL,
-                                                                    RestAPIPublisherImpl.username,
-                                                                    RestAPIPublisherImpl.password,
-                                                                    APIMIntegrationConstants.SUPER_TENANT_DOMAIN);
-        ClientAuthenticator.makeDCRRequest(publisherParamRequest);
-
-        //DCR call for dev portal app
-        DCRParamRequest devPortalParamRequest = new DCRParamRequest(RestAPIStoreImpl.appName,
-                                                                    RestAPIStoreImpl.callBackURL,
-                                                                    RestAPIStoreImpl.tokenScope,
-                                                                    RestAPIStoreImpl.appOwner,
-                                                                    RestAPIStoreImpl.grantType, dcrURL,
-                                                                    RestAPIStoreImpl.username,
-                                                                    RestAPIStoreImpl.password,
-                                                                    APIMIntegrationConstants.SUPER_TENANT_DOMAIN);
-        ClientAuthenticator.makeDCRRequest(devPortalParamRequest);
-
-        // DCR call for admin portal app
-        DCRParamRequest adminPortalParamRequest = new DCRParamRequest(RestAPIAdminImpl.appName,
-                                                                      RestAPIAdminImpl.callBackURL,
-                                                                      RestAPIAdminImpl.tokenScope,
-                                                                      RestAPIAdminImpl.appOwner,
-                                                                      RestAPIAdminImpl.grantType, dcrURL,
-                                                                      RestAPIAdminImpl.username,
-                                                                      RestAPIAdminImpl.password,
-                                                                      APIMIntegrationConstants.SUPER_TENANT_DOMAIN);
-        ClientAuthenticator.makeDCRRequest(adminPortalParamRequest);
     }
 
-    @AfterSuite(description = "stop the setup")
+    @AfterSuite(description = "stop API Manager")
     public void stopAPIM() {
         apimInstance.stopAPIM();
     }
@@ -112,20 +61,5 @@ public class WithApimBeforeTestSuite extends ApimAdvancedBaseTest {
     private Boolean checkForAPIMServerStartup() throws IOException {
         HttpResponse response = HttpClientRequest.doGet(Utils.getAPIMServiceURLHttp("/services/Version"));
         return Objects.nonNull(response) && response.getResponseCode() == 200;
-    }
-
-    /**
-     * Helper method to set the SSL context.
-     */
-    protected void setSSlSystemProperties() {
-        URL certificatesTrustStore = getClass().getClassLoader()
-                .getResource("keystore/client-truststore.jks");
-        if (certificatesTrustStore != null) {
-            System.setProperty("javax.net.ssl.trustStore", certificatesTrustStore.getPath());
-        } else {
-            log.error("Truststore is not set.");
-        }
-        System.setProperty("javax.net.ssl.trustStorePassword", "wso2carbon");
-        System.setProperty("https.protocols", "TLSv1,TLSv1.1,TLSv1.2");
     }
 }
